+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 - Allow `o` keybinding from the lucky prompt.
+- Permanent help text ([#24](https://github.com/samtay/so/issues/24)).
 
 #### Fixed
 - Hardcoded stackoverflow link