+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.9]
+
 #### Added
 - Allow `o` keybinding from the lucky prompt.
 - Permanent help text ([#24](https://github.com/samtay/so/issues/24)).
@@ -85,7 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.5]
 - (Unofficial) initial passable release
 
-[Unreleased]: (https://github.com/samtay/so/compare/v0.4.8...HEAD)
+[Unreleased]: (https://github.com/samtay/so/compare/v0.4.9...HEAD)
+[0.4.9]: (https://github.com/samtay/so/compare/v0.4.8...v0.4.9)
 [0.4.8]: (https://github.com/samtay/so/compare/v0.4.7...v0.4.8)
 [0.4.7]: (https://github.com/samtay/so/compare/v0.4.6...v0.4.7)
 [0.4.6]: (https://github.com/samtay/so/compare/v0.4.5...v0.4.6)